+++ packages/calls/ios/salesiq_mobilisten_calls/Package.swift
@@ -2,12 +2,12 @@
 import PackageDescription
 
 let package = Package(
-    name: "salesiq_mobilisten",
+    name: "salesiq_mobilisten_calls",
     platforms: [
         .iOS("13.0")
     ],
     products: [
-        .library(name: "salesiq-mobilisten", targets: ["salesiq_mobilisten"])
+        .library(name: "salesiq-mobilisten-calls", targets: ["salesiq_mobilisten_calls"])
     ],
     dependencies: [
         .package(url: "https://github.com/zoho/SalesIQ-Mobilisten-iOS-SP.git", exact: "10.4.8"),
@@ -15,9 +15,10 @@ let package = Package(
     ],
     targets: [
         .target(
-            name: "salesiq_mobilisten",
+            name: "salesiq_mobilisten_calls",
             dependencies: [
                 .product(name: "Mobilisten", package: "SalesIQ-Mobilisten-iOS-SP"),
+                .product(name: "MobilistenCalls", package: "SalesIQ-Mobilisten-iOS-SP"),
                 .product(name: "FlutterFramework", package: "FlutterFramework")
             ]
         )