+++ packages/calls/ios/salesiq_mobilisten_calls/Sources/salesiq_mobilisten_calls/MobilistenCallsPlugin.swift
@@ -1,0 +1,7 @@
+import Flutter
+
+public final class MobilistenCallsPlugin: NSObject, FlutterPlugin {
+    public static func register(with registrar: FlutterPluginRegistrar) {
+        SwiftMobilistenCallsPlugin.register(with: registrar)
+    }
+}